debt - remove referenceSearch/common
